refactor/DataLoading/requirements.txt
async-timeout==4.0.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dnspython==2.3.0
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pymongo==4.3.3
python-dotenv==1.0.0
redis==4.5.4
requests==2.30.0
urllib3==2.0.2
Werkzeug==2.3.3
